+++ src/IQFeed.CSharpApiClient/Lookup/Historical/Messages/TickMessage.cs
@@ -10,7 +10,7 @@ namespace IQFeed.CSharpApiClient.Lookup.Historical.Messages
     {
         public const string TickDateTimeFormat = "yyyy-MM-dd HH:mm:ss.ffffff";
 
-        public TickMessage(DateTime timestamp, double last, long lastSize, long totalVolume, double bid, double ask,
+        public TickMessage(DateTime timestamp, double last, int lastSize, long totalVolume, double bid, double ask,
             long tickId, char basisForLast, int tradeMarketCenter, string tradeConditions, string requestId = null)
         {
             Timestamp = timestamp;
@@ -28,7 +28,7 @@ namespace IQFeed.CSharpApiClient.Lookup.Historical.Messages
 
         public DateTime Timestamp { get; private set; }
         public double Last { get; private set; }
-        public long LastSize { get; private set; }
+        public int LastSize { get; private set; }
         public long TotalVolume { get; private set; }
         public double Bid { get; private set; }
         public double Ask { get; private set; }
@@ -45,7 +45,7 @@ namespace IQFeed.CSharpApiClient.Lookup.Historical.Messages
             return new TickMessage(
                 timestamp: DateTime.ParseExact(s: values[0], format: TickDateTimeFormat, CultureInfo.InvariantCulture),
                 last: double.Parse(values[1], CultureInfo.InvariantCulture),
-                lastSize: long.Parse(values[2], CultureInfo.InvariantCulture),
+                lastSize: int.Parse(values[2], CultureInfo.InvariantCulture),
                 totalVolume: long.Parse(values[3], CultureInfo.InvariantCulture),
                 bid: double.Parse(values[4], CultureInfo.InvariantCulture),
                 ask: double.Parse(values[5], CultureInfo.InvariantCulture),
@@ -63,7 +63,7 @@ namespace IQFeed.CSharpApiClient.Lookup.Historical.Messages
             return new TickMessage(
                 timestamp: DateTime.ParseExact(values[1], TickDateTimeFormat, CultureInfo.InvariantCulture),
                 last: double.Parse(values[2], CultureInfo.InvariantCulture),
-                lastSize: long.Parse(values[3], CultureInfo.InvariantCulture),
+                lastSize: int.Parse(values[3], CultureInfo.InvariantCulture),
                 totalVolume: long.Parse(values[4], CultureInfo.InvariantCulture),
                 bid: double.Parse(values[5], CultureInfo.InvariantCulture),
                 ask: double.Parse(values[6], CultureInfo.InvariantCulture),
